refactor: Move AgreementChangesAlert component to new directory
The AgreementChangesAlert component has been moved from the "UI/Alert" directory to the "Agreements" directory. This reorganization improves the code structure and makes it easier to locate and manage components related to agreements.

Related User Commits:
- feat: adds AgreementChangesAlert to Agreement details pages
- feat: Add isClosable prop to AgreementChangesAlert component
