+++ frontend/src/components/Agreements/AgreementChangesAlert/index.js
@@ -1,0 +1,2 @@
+export * from "./AgreementsChangesAlert.jsx";
+export { default } from "./AgreementsChangesAlert.jsx";
